Divide out digraph and undirected graph into interface files, start on global nodes
